appraisal-automation/requirements.txt
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
python-docx>=1.1.0
pydantic>=2.0.0
lxml>=5.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
